Old obj folder deleted + include folder only contains one file
