+++ RichardEngine/include/richard.h
@@ -1,0 +1,5 @@
+/*
+* This header contains all the headers that will be included in RichardEditor
+*/
+
+#include "engine.h"
